+++ change-log.md
@@ -239,4 +239,5 @@
 
 ##### Version 0.5.5
 - Remove unnecessary packages from devDependencies
+- Rename `CreateEditModal` & `CreateEditModalProvider` to `FormModal` & `FormModalProvider`
 - Remove out of context error when clicking row to open `EditModal` on `DataTable`